added contents in lexer.c; added lexer.c in main
